# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TestRunStatus = typeof TestRunStatus[keyof typeof TestRunStatus];


export const TestRunStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface TestRun {
  id: string;
  url: string;
  /** @nullable */
  environmentTag?: string | null;
  status: TestRunStatus;
  moduleSelection: string[];
  /** @nullable */
  triggeredBy?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  /** @nullable */
  totalScenarios?: number | null;
  /** @nullable */
  passedScenarios?: number | null;
  /** @nullable */
  failedScenarios?: number | null;
  /** @nullable */
  passRate?: number | null;
  /** @nullable */
  durationMs?: number | null;
}

export interface TestRunInput {
  /** @minLength 1 */
  url: string;
  environmentTag?: string;
  /** @minItems 1 */
  moduleSelection: string[];
  testUsername?: string;
  testPassword?: string;
}

export type TestResultModule = typeof TestResultModule[keyof typeof TestResultModule];


export const TestResultModule = {
  home: 'home',
  navigation: 'navigation',
  search: 'search',
  category: 'category',
  plp: 'plp',
  pdp: 'pdp',
  cart: 'cart',
  checkout: 'checkout',
} as const;

export type TestResultStatus = typeof TestResultStatus[keyof typeof TestResultStatus];


export const TestResultStatus = {
  pass: 'pass',
  fail: 'fail',
  warning: 'warning',
  skipped: 'skipped',
} as const;

export interface TestResult {
  id: string;
  runId: string;
  module: TestResultModule;
  scenario: string;
  testId: string;
  status: TestResultStatus;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  screenshotUrl?: string | null;
  apisCalled?: string[];
  /** @nullable */
  productName?: string | null;
  /** @nullable */
  productCode?: string | null;
  createdAt: string;
}

export type PerformanceMetricPageType = typeof PerformanceMetricPageType[keyof typeof PerformanceMetricPageType];


export const PerformanceMetricPageType = {
  home: 'home',
  plp: 'plp',
  pdp: 'pdp',
} as const;

export interface PerformanceMetric {
  id: string;
  runId: string;
  pageType: PerformanceMetricPageType;
  /** @nullable */
  ttfbMs?: number | null;
  /** @nullable */
  fcpMs?: number | null;
  /** @nullable */
  lcpMs?: number | null;
  /** @nullable */
  loadTimeMs?: number | null;
  /** @nullable */
  pageWeightKb?: number | null;
  /** @nullable */
  networkRequests?: number | null;
  createdAt: string;
}

export interface ModuleBreakdown {
  module: string;
  passRate: number;
  totalScenarios: number;
  passedScenarios: number;
  failedScenarios: number;
}

export type PerformanceSnapshotPageType = typeof PerformanceSnapshotPageType[keyof typeof PerformanceSnapshotPageType];


export const PerformanceSnapshotPageType = {
  home: 'home',
  plp: 'plp',
  pdp: 'pdp',
} as const;

export interface PerformanceSnapshot {
  pageType: PerformanceSnapshotPageType;
  /** @nullable */
  avgLoadTimeMs?: number | null;
  /** @nullable */
  avgLcpMs?: number | null;
  /** @nullable */
  avgTtfbMs?: number | null;
  threshold: number;
  breachCount: number;
}

export interface DashboardSummary {
  totalRuns: number;
  totalRunsThisWeek: number;
  overallPassRate: number;
  /** @nullable */
  averageDurationMs?: number | null;
  recentRuns: TestRun[];
  moduleBreakdown: ModuleBreakdown[];
  performanceSnapshot: PerformanceSnapshot[];
}

export type PerformanceTrendPageType = typeof PerformanceTrendPageType[keyof typeof PerformanceTrendPageType];


export const PerformanceTrendPageType = {
  home: 'home',
  plp: 'plp',
  pdp: 'pdp',
} as const;

export interface PerformanceTrend {
  runId: string;
  runAt: string;
  pageType: PerformanceTrendPageType;
  /** @nullable */
  ttfbMs?: number | null;
  /** @nullable */
  fcpMs?: number | null;
  /** @nullable */
  lcpMs?: number | null;
  /** @nullable */
  loadTimeMs?: number | null;
  url: string;
}

export interface RunComparison {
  runA: TestRun;
  runB: TestRun;
  resultsA: TestResult[];
  resultsB: TestResult[];
  metricsA: PerformanceMetric[];
  metricsB: PerformanceMetric[];
  improvements: string[];
  regressions: string[];
}

export type ListRunsParams = {
status?: ListRunsStatus;
environment?: string;
limit?: number;
offset?: number;
};

export type ListRunsStatus = typeof ListRunsStatus[keyof typeof ListRunsStatus];


export const ListRunsStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export type GetPerformanceTrendsParams = {
url?: string;
pageType?: GetPerformanceTrendsPageType;
limit?: number;
};

export type GetPerformanceTrendsPageType = typeof GetPerformanceTrendsPageType[keyof typeof GetPerformanceTrendsPageType];


export const GetPerformanceTrendsPageType = {
  home: 'home',
  plp: 'plp',
  pdp: 'pdp',
} as const;

export type CompareRunsParams = {
runA: string;
runB: string;
};
