# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleBreakdown } from './moduleBreakdown';
import type { PerformanceSnapshot } from './performanceSnapshot';
import type { TestRun } from './testRun';

export interface DashboardSummary {
  totalRuns: number;
  totalRunsThisWeek: number;
  overallPassRate: number;
  /** @nullable */
  averageDurationMs?: number | null;
  recentRuns: TestRun[];
  moduleBreakdown: ModuleBreakdown[];
  performanceSnapshot: PerformanceSnapshot[];
}