# lib/api-zod/src/generated/types/moduleBreakdown.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ModuleBreakdown {
  module: string;
  passRate: number;
  totalScenarios: number;
  passedScenarios: number;
  failedScenarios: number;
}