# lib/api-zod/src/generated/types/performanceMetric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PerformanceMetricPageType } from './performanceMetricPageType';

export interface PerformanceMetric {
  id: string;
  runId: string;
  pageType: PerformanceMetricPageType;
  /** @nullable */
  ttfbMs?: number | null;
  /** @nullable */
  fcpMs?: number | null;
  /** @nullable */
  lcpMs?: number | null;
  /** @nullable */
  loadTimeMs?: number | null;
  /** @nullable */
  pageWeightKb?: number | null;
  /** @nullable */
  networkRequests?: number | null;
  createdAt: string;
}