# lib/api-zod/src/generated/types/runComparison.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PerformanceMetric } from './performanceMetric';
import type { TestResult } from './testResult';
import type { TestRun } from './testRun';

export interface RunComparison {
  runA: TestRun;
  runB: TestRun;
  resultsA: TestResult[];
  resultsB: TestResult[];
  metricsA: PerformanceMetric[];
  metricsB: PerformanceMetric[];
  improvements: string[];
  regressions: string[];
}