# lib/api-zod/src/generated/types/testRunInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * B2C Storefront Test Automation Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TestRunInput {
  /** @minLength 1 */
  url: string;
  environmentTag?: string;
  /** @minItems 1 */
  moduleSelection: string[];
  testUsername?: string;
  testPassword?: string;
}